move log density helpers to main directory
