+++ pyloo/tests/base_tests/test_loo_approximate_posterior.py
@@ -6,10 +6,10 @@ import pytest
 import xarray as xr
 from arviz.data import InferenceData
 
+from ...log_density import compute_log_weights
 from ...loo import loo
 from ...loo_approximate_posterior import loo_approximate_posterior
 from ...wrapper.laplace import Laplace
-from ...wrapper.log_density import compute_log_weights
 from ..helpers import assert_arrays_allclose
